requirements.txt
asgiref==3.4.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2
django-allauth==0.47.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
elasticsearch==7.16.1
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
locust==2.5.1
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.21.4
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
psutil==5.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
roundrobin==0.0.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
yml==0.0.1
zope.event==4.5.0
zope.interface==5.4.0
